fix: address lint findings carried over from previous theme polish work
Pre-push linting flagged a handful of issues introduced by the recent
theme-polish commits (regex backtracking risk in the frontmatter-H1 plugin,
unsafe type assertions in the callout plugin, a window→globalThis swap in
the error boundary, and Search.tsx tipping over the 300-line file limit).

Resolved without behavioural changes:
- remark-strip-frontmatter-h1: regex made linear (\\S after [ \\t]+),
  prefer-replaceAll, removed the unsafe Heading cast in favour of a type
  narrow on the array element.
- remark-callouts: typed the array-index access as `| undefined` so the
  runtime guards are honest to the lint rules instead of being flagged
  as unnecessary.
- ErrorBoundary: window.location → globalThis.location.
- Search.tsx: extracted the two utility hooks (global ⌘K shortcut and
  outside-click handler) into a new search-hooks.ts module, bringing
  Search.tsx back below the 300-line limit.
