+++ packages/ardo/src/ui/components/Search.tsx
@@ -1,10 +1,11 @@
 import MiniSearch from "minisearch"
-import { useEffect, useMemo, useRef, useState } from "react"
+import { useMemo, useRef, useState } from "react"
 import { Link, useNavigate } from "react-router"
 import searchDocs from "virtual:ardo/search-index"
 
 import { SearchIcon } from "../icons"
 import { ArdoOwlMark } from "../OwlMark"
+import { useGlobalSearchShortcut, useOutsideClick } from "./search-hooks"
 import * as styles from "./Search.css"
 import { SearchPopover } from "./SearchPopover"
 
@@ -26,59 +27,6 @@ type SearchMatch = {
 export type ArdoSearchProps = {
   /** Placeholder text for the search input (default: "Search...") */
   placeholder?: string
-}
-
-function useGlobalSearchShortcut(
-  inputRef: React.RefObject<HTMLInputElement | null>,
-  setIsOpen: (open: boolean) => void
-) {
-  useEffect(() => {
-    const handleGlobalKeyDown = (e: KeyboardEvent) => {
-      if ((e.metaKey || e.ctrlKey) && e.key === "k") {
-        e.preventDefault()
-        inputRef.current?.focus()
-        setIsOpen(true)
-      }
-      if (e.key === "Escape") {
-        setIsOpen(false)
-      }
-    }
-    document.addEventListener("keydown", handleGlobalKeyDown)
-    return () => {
-      document.removeEventListener("keydown", handleGlobalKeyDown)
-    }
-  }, [inputRef, setIsOpen])
-}
-
-type OutsideClickOptions = {
-  containerRef: React.RefObject<HTMLDivElement | null>
-  popoverClass: string
-  isOpen: boolean
-  setIsOpen: (open: boolean) => void
-}
-
-function useOutsideClick({ containerRef, popoverClass, isOpen, setIsOpen }: OutsideClickOptions) {
-  useEffect(() => {
-    if (!isOpen) return
-    const handleOutsideClick = (e: MouseEvent | TouchEvent) => {
-      if (!(e.target instanceof Element)) {
-        return
-      }
-
-      const target = e.target
-      const inContainer = containerRef.current?.contains(target) === true
-      const inPopover = target.closest(`.${popoverClass}`) != null
-      if (!inContainer && !inPopover) {
-        setIsOpen(false)
-      }
-    }
-    document.addEventListener("mousedown", handleOutsideClick)
-    document.addEventListener("touchstart", handleOutsideClick)
-    return () => {
-      document.removeEventListener("mousedown", handleOutsideClick)
-      document.removeEventListener("touchstart", handleOutsideClick)
-    }
-  }, [containerRef, popoverClass, isOpen, setIsOpen])
 }
 
 function useSearchIndex() {
